Moved StringUtil to the appropriate package
