+++ commons/src/main/java/org/eclipse/kapua/commons/util/StringUtil.java
@@ -9,14 +9,14 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.commons.configuration;
-
-import java.util.ArrayList;
-import java.util.List;
+package org.eclipse.kapua.commons.util;
 
 import org.eclipse.kapua.KapuaException;
 import org.eclipse.kapua.commons.configuration.metatype.Password;
 import org.eclipse.kapua.commons.configuration.metatype.TscalarImpl;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Utilities to manipulate string
@@ -63,79 +63,78 @@ public class StringUtil {
         for (int i = 0; i < strValues.length(); i++) {
             char c1 = strValues.charAt(i);
             switch (c1) {
-            case DELIMITER:
-                // When the delimiter is encountered, add the extracted
-                // token to the result and prepare the buffer to receive the
-                // next token.
-                values.add(buffer.toString());
-                buffer.delete(0, buffer.length());
-                break;
-            case ESCAPE:
-                // When the escape is encountered, add the immediately
-                // following character to the token, unless the end of the
-                // input has been reached. Note this will result in loop
-                // counter 'i' being incremented twice, once here and once
-                // at the end of the loop.
-                if (i + 1 < strValues.length()) {
-                    buffer.append(strValues.charAt(++i));
-                }
-                // If the ESCAPE character occurs as the last character
-                // of the string, ignore it.
-                break;
-            default:
-                // For all other characters, add them to the current token
-                // unless dealing with unescaped whitespace at the beginning
-                // or end. We know the whitespace is unescaped because it
-                // would have been handled in the ESCAPE case otherwise.
-                if (Character.isWhitespace(c1)) {
-                    // Ignore unescaped whitespace at the beginning of the
-                    // token.
-                    if (buffer.length() == 0) {
-                        continue;
+                case DELIMITER:
+                    // When the delimiter is encountered, add the extracted
+                    // token to the result and prepare the buffer to receive the
+                    // next token.
+                    values.add(buffer.toString());
+                    buffer.delete(0, buffer.length());
+                    break;
+                case ESCAPE:
+                    // When the escape is encountered, add the immediately
+                    // following character to the token, unless the end of the
+                    // input has been reached. Note this will result in loop
+                    // counter 'i' being incremented twice, once here and once
+                    // at the end of the loop.
+                    if (i + 1 < strValues.length()) {
+                        buffer.append(strValues.charAt(++i));
                     }
-                    // If the whitespace is not at the beginning, look
-                    // forward, starting with the next character, to see if
-                    // it's in the middle or at the end. Unescaped
-                    // whitespace in the middle is okay.
-                    for (int j = i + 1; j < strValues.length(); j++) {
-                        // Keep looping until the end of the string is
-                        // reached or a non-whitespace character other than
-                        // the escape is seen.
-                        char c2 = strValues.charAt(j);
-                        if (!Character.isWhitespace(c2)) {
-                            // If the current character is not the DELIMITER, all whitespace
-                            // characters are significant and should be added to the token.
-                            // Otherwise, they're at the end and should be ignored. But watch
-                            // out for an escape character at the end of the input. Ignore it
-                            // and any previous insignificant whitespace if it exists.
-                            if (c2 == ESCAPE && j + 1 >= strValues.length()) {
-                                continue;
-                            }
-                            if (c2 != DELIMITER) {
-                                buffer.append(strValues.substring(i, j));
-                            }
-                            // Let loop counter i catch up with the inner loop but keep in
-                            // mind it will still be incremented at the end of the outer loop.
-                            i = j - 1;
-                            break;
+                    // If the ESCAPE character occurs as the last character
+                    // of the string, ignore it.
+                    break;
+                default:
+                    // For all other characters, add them to the current token
+                    // unless dealing with unescaped whitespace at the beginning
+                    // or end. We know the whitespace is unescaped because it
+                    // would have been handled in the ESCAPE case otherwise.
+                    if (Character.isWhitespace(c1)) {
+                        // Ignore unescaped whitespace at the beginning of the
+                        // token.
+                        if (buffer.length() == 0) {
+                            continue;
                         }
+                        // If the whitespace is not at the beginning, look
+                        // forward, starting with the next character, to see if
+                        // it's in the middle or at the end. Unescaped
+                        // whitespace in the middle is okay.
+                        for (int j = i + 1; j < strValues.length(); j++) {
+                            // Keep looping until the end of the string is
+                            // reached or a non-whitespace character other than
+                            // the escape is seen.
+                            char c2 = strValues.charAt(j);
+                            if (!Character.isWhitespace(c2)) {
+                                // If the current character is not the DELIMITER, all whitespace
+                                // characters are significant and should be added to the token.
+                                // Otherwise, they're at the end and should be ignored. But watch
+                                // out for an escape character at the end of the input. Ignore it
+                                // and any previous insignificant whitespace if it exists.
+                                if (c2 == ESCAPE && j + 1 >= strValues.length()) {
+                                    continue;
+                                }
+                                if (c2 != DELIMITER) {
+                                    buffer.append(strValues.substring(i, j));
+                                }
+                                // Let loop counter i catch up with the inner loop but keep in
+                                // mind it will still be incremented at the end of the outer loop.
+                                i = j - 1;
+                                break;
+                            }
+                        }
+                    } else {
+                        // For non-whitespace characters.
+                        buffer.append(c1);
                     }
-                } else {
-                    // For non-whitespace characters.
-                    buffer.append(c1);
-                }
             }
         }
         // Don't forget to add the last token.
         values.add(buffer.toString());
-        return values.toArray(new String[] {});
+        return values.toArray(new String[]{});
     }
 
     /**
      * Convert the string to the appropriate Object based on type
      *
-     * @param type
-     *            allowed values are {@link TscalarImpl}
+     * @param type   allowed values are {@link TscalarImpl}
      * @param string
      * @return
      * @throws KapuaException
