+++ commons/src/test/java/org/eclipse/kapua/commons/configuration/StringUtilTest.java
@@ -11,9 +11,9 @@
  *******************************************************************************/
 package org.eclipse.kapua.commons.configuration;
 
-import org.eclipse.kapua.commons.configuration.metatype.Password;
-
 import org.assertj.core.api.Assertions;
+import org.eclipse.kapua.commons.configuration.metatype.Password;
+import org.eclipse.kapua.commons.util.StringUtil;
 import org.eclipse.kapua.qa.markers.junit.JUnitTests;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -24,7 +24,7 @@ public class StringUtilTest {
     @Test
     public void shouldConvertPasswordsArrayToString() {
         // Given
-        Password[] passwords = new Password[] {new Password("foo") };
+        Password[] passwords = new Password[]{new Password("foo")};
 
         // When
         String passwordsString = StringUtil.valueToString(passwords);
